requirements.txt
# Python 3.11 ONLY — KHÔNG dùng 3.12+
# py -3.11 -m pip install -r requirements.txt

# === AI Core ===
anthropic==0.84.0
langgraph>=0.2.0
langchain-anthropic>=0.2.0
langchain-core>=0.2.0

# === Data ===
vnstock==3.4.2
yfinance>=0.2.40
pandas>=2.0.0,<3.0.0
numpy>=1.24.0,<2.0.0

# === Technical Indicators ===
pandas-ta-openbb>=0.4.0

# === Web Scraping ===
beautifulsoup4>=4.12.0
requests>=2.31.0

# === Discord ===
discord.py>=2.3.0

# === Dashboard ===
streamlit>=1.35.0
plotly>=5.20.0

# === Scheduler ===
apscheduler>=3.10.0

# === Utilities ===
python-dotenv>=1.0.0
tenacity>=8.2.0

# === Memory (Phase 6) ===
# chromadb>=0.5.0

# === Observability ===
langsmith>=0.1.0
